requirements.txt
requests==2.31.0
smtplib
email
python-dotenv==1.0.0
schedule==1.2.0